Auth and login type fixes.
